requirements.txt
azure-mgmt-containerservice
azure-identity
python-dotenv
kubernetes
